[BE] refactor: CategoryBookmarkCount 경로 projection 패키지로 변경 #774
