+++ backend/app/src/main/java/com/onair/hearit/app/explore/application/scorefactor/BookmarkScoreFactor.java
@@ -5,8 +5,8 @@ import com.onair.hearit.core.domain.Hearit;
 import com.onair.hearit.core.domain.Member;
 import com.onair.hearit.core.domain.UserType;
 import com.onair.hearit.core.infrastructure.jpa.BookmarkRepository;
-import com.onair.hearit.core.infrastructure.jpa.CategoryBookmarkCount;
 import com.onair.hearit.core.infrastructure.jpa.MemberRepository;
+import com.onair.hearit.core.infrastructure.projection.CategoryBookmarkCount;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
