+++ backend/core/src/main/java/com/onair/hearit/core/infrastructure/projection/CategoryBookmarkCount.java
@@ -1,4 +1,4 @@
-package com.onair.hearit.core.infrastructure.jpa;
+package com.onair.hearit.core.infrastructure.projection;
 
 public interface CategoryBookmarkCount {
